backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
python-dotenv
python-multipart
aiofiles
pdfplumber
python-docx
python-pptx
pillow
pytesseract
pydantic-settings
google-cloud-storage
google-cloud-vision
google-generativeai
google-adk
httpx
